### 2019-09-24 #1760 @whedon
[PRE REVIEW]: pymccrgb: Color- and curvature-based classification of multispectral point clouds in Python
**Submitting author:** @rmsare (<a href="http://orcid.org/0000-0003-3711-6771">Robert Sare</a>)
**Repository:** <a href="http://github.com/rmsare/pymccrgb" target ="_blank">http://github.com/rmsare/pymccrgb</a>
**Version:** v0.1.0
