### 2019-09-24 #1760 @danielskatz
@rmsare  - it looks like the repo you entered doesn't exist - but it's just that the http:// was missing - I've added it now
@arfon - does this need to be changed in an internal database as well?
