### 2019-09-24 #1760 @danielskatz
@rmsare - can you accept https://github.com/rmsare/pymccrgb/pull/14 and then enter a new comment in this issue ```@whedon generate pdf``` to see if that fixes the problem
