### 2019-09-24 #1760 @rmsare
Ack, thank you @danielskatz for fixing these issues.

I'd like to note that one of the references -- Pedregosa et al, 2011 -- does not have a DOI. The [original JMLR article](http://www.jmlr.org/papers/volume12/pedregosa11a/pedregosa11a.pdf) and [arxiv version](https://ui.adsabs.harvard.edu/abs/2012arXiv1201.0490P/abstract) do not have DOIs and the article isn't available in CrossRef as far as I can tell.
