### 2019-09-24 bfb5e08
Condense readme instructions
