### 2019-09-25 45ee172
Update install instructions
