### 2019-09-25 c215fe8
Change install instructions - uplaoded to PyPI
