+++ docs/source/installation.rst
@@ -1,7 +1,7 @@
 Installation
 ============
 
-``pymccrgb`` is hosted on GitHub. You can install it with ``conda`` or 
+``pymccrgb`` is hosted on GitHub and PyPI. You can install it with ``conda`` or 
 ``virtualenv`` in a virtual environment.
 
 .. code-block:: bash
@@ -10,7 +10,7 @@ Installation
     cd pymccrgb
     conda env create -f environment.yml
     conda activate pymcc
-    pip install .
+    pip install pymccrgb
     py.test pymccrgb/tests
 
 The package requires several C dependencies that may not be installed on your
