### 2019-09-25 9617162
Add manifest for sample data
