+++ MANIFEST.in
@@ -1,3 +1,1 @@
 include pymccrgb/data/*.laz
-include pymccrgb/tests/data/*.laz
-include pymccrgb/tests/output/*.npy
