### 2019-09-26 #1760 @kthyng
Thanks!

Hi @margauxmouchene and @martibosch! Are you interested and able to review this JOSS submission in about the next three weeks? You can find more details [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
