### 2019-09-30 d87e070
Improve install description
