### 2019-09-30 b537a7a
Shorten install description
