### 2019-09-30 93fc22e
More readme improvements
