+++ paper/paper.bib
@@ -36,9 +36,8 @@ ISSN={1521-9615},
 month={March},}
 
 @misc{pdal2018,
-author = {PDAL Contributors},
+author = {{PDAL Contributors}},
 title = {PDAL Point Data Abstraction Library},
-month = nov, 
 year = 2018, 
 doi = {10.5281/zenodo.2556738}, 
 url = {https://doi.org/10.5281/zenodo.2556738}
