### 2019-10-01 a1d1f98
Add test data to manifest
