+++ MANIFEST.in
@@ -1,2 +1,3 @@
-recursive-include pymccrgb/data/ *.laz
+recursive-include pymccrgb/data *.laz
+recursive-include pymccrgb/tests/data *.laz
 recursive-include pymccrgb/tests *.py
