### 2019-10-01 364f347
Add generic pdal output function and write_las
