### 2019-10-02 #1760 @kthyng
Hi @callumrollo, @daniellivingston, and @banesullivan! I need one more reviewer for this JOSS submission. Are any of you interested in reviewing? We ask for reviews within about 3 weeks and you can get more information about the process and details here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
