### 2019-10-02 #1760 @daniellivingston
Hi @kthyng , I would be more than happy to review this submission.
