### 2019-10-02 #1760 @kthyng
@daniellivingston perfect! We can proceed with this review then.
