### 2019-10-02 #1760 @banesullivan
Hi @kthyng - looks like @daniellivingston beat me here! This library looks really cool and I'd be happy to review if needed (although a bit busy lately). Thanks for thinking of me! I'll definitely be scoping out this package and maybe finding a way to couple [PyVista](https://docs.pyvista.org) with it
