### 2019-10-02 #1760 @kthyng
@martibosch and @daniellivingston head on over to #1777 to do your reviews.

@banesullivan Thanks for being willing to help out! I'll keep you in mind for the next relevant submission.
