### 2019-10-02 #1777 @kthyng
I just edited the first comment in this issue to update all of the URLs (I think). Let me know if you find one with the old one.
