### 2019-10-02 #1777 @rmsare
I wanted to note that one reference ([Pedregosa, et al., 2011, JMLR](http://www.jmlr.org/papers/volume12/pedregosa11a/pedregosa11a.pdf)) does not have a DOI.

It appears that the journal did not assign it a DOI as it's not in CrossRef. The arxiv version of this paper has the same issue.
