### 2019-10-04 64a93c2
Add argument to sample arbitrary indices of point cloud
