+++ pymccrgb/pointutils.py
@@ -2,6 +2,7 @@
 
 import fiona
 import numpy as np
+import pdal
 import subprocess
 
 from scipy.spatial import cKDTree
@@ -62,14 +63,31 @@ def point_grid(x_min, x_max, y_min, y_max, dx, dy=None):
     return points
 
 
-def sample_point_cloud(source, target):
+def sample_point_cloud(source, target, sample_indices=[2]):
     """ Resamples a source point cloud at the coordinates of a target points
 
         Uses the nearest point in the target point cloud to the source point
+
+    Parameters
+    ----------
+        source: array
+            Input point cloud
+
+        target: array
+            Target point cloud for sample locations
+        
+        sample_indices: list
+            List of indices to sample from source. Defaults to 2 (z or height
+            dimension)
+
+    Returns
+    -------
+        An array of sampled points
     """
+    sample_indices = np.array(sample_indices)
     tree = cKDTree(source[:, 0:2])
     dist, idx = tree.query(target, n_jobs=-1)
-    output = np.hstack([target, source[idx, 2].reshape((len(idx), 1))])
+    output = np.hstack([target, source[idx[:, None], sample_indices].reshape((len(idx), len(sample_indices)))])
     return output
 
 
@@ -140,3 +158,7 @@ def stratified_sample(X, y, size=100, seed=None):
     y_sampled = np.vstack(ys)
 
     return X_sampled, y_sampled
+
+
+def write_dem(data, filename, resolution=1, radius=None):
+    pass
