### 2019-10-04 57e4f19
Add write_pdal function that uses temp text file (:()
