### 2019-10-04 c48cac6
Add write_dem convenience function
