### 2019-10-04 b4cdad4
Add default radius
