### 2019-10-09 d6f2b23
Delete repeated word
