### 2019-10-09 #1777 @daniellivingston
@rmsare , on x86_64 Ubuntu, running the command:

```
$ sudo apt-get install -y liblas-c-dev
```

gives the error:

```
404 Not Found

E: Aborting install.
```

Trying to run `conda env create -f environment.yml` fails due to `liblas/writer.hpp` not being found.
