### 2019-10-09 #1777 @daniellivingston
@rmsare , I have completed my review. Great package that may find a use within my research group. I am recommending your publication for review.

On your ReadTheDocs documentation, the left sidebar has an area where the background is pure white, obstructing the package name. This is a minor thing and does not impact legibility of the documentation, but should be addressed.
