### 2019-10-09 66283dc
Try to update header color for RTD
