+++ docs/source/conf.py
@@ -162,7 +162,7 @@ html_theme_path = [msmb_theme.get_html_theme_path()]
 # further.  For a list of options available for each theme, see the
 # documentation.
 #
-# html_theme_options = {}
+html_theme_options = {"style_nav_header_background": "#2980B9"}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
