### 2019-10-09 #1777 @rmsare
@daniellivingston Thanks. Unfortunately the installation is a little finicky, but I'm glad it worked in your VM.

I've fixed the header color so the title is now readable on https://pymccrgb.readthedocs.io/en/latest/
