### 2019-10-11 d97450d
Fix training_tol in example
