### 2019-10-18 #1777 @rmsare
@martibosch Thanks for your detailed review. Just to give myself a TODO list, here is what I think needs to be addressed:

- [ ] Improve installation, possibly with conda recipe
- [ ] Remove example data from package, but download it when `datasets.load_...` is first called
- [ ] Fix `flake8` failures and run a linter
- [ ] (Optional?) Use Coveralls CI with tests

Please let me know if there's anything I've missed. I will work on these soon and ping you all in this thread when I'm done.

I'm glad to hear you think this package will be useful to your work, too!
