### 2019-10-18 #1777 @rmsare
... and to clarify authorship, the second author @gehilley has supported development and some related data acquisition. (He is my PhD supervisor.)

Code-wise, he is also responsible for the wrapper this package depends on, which is hosted separately at https://github.com/stgl/pymcc
