### 2019-10-18 rmsare/pymccrgb#17 @rmsare
JOSS review: Improve installation
Installation: while I could successfully install the package by following the instructions of the documentation, I believe that the procedure is a bit too convoluted, since it requires cloning the source repository just for the environment file and then suggests to install the package from PyPI when, given that the repository has already been cloned, the package could be installed as in `pip install .`. I would strongly suggest the authors to create a conda recipe so that the package con be installed from conda-forge without the need for cloning the source repository.

openjournals/joss-reviews#1777
