### 2019-10-18 rmsare/pymccrgb#18 @rmsare
JOSS review: Download and cache datasets
Finally, I wonder whether it is essential to include a 21.6 MB example data which practically accounts for the totality of the size of the Python wheel or source tar. I believe that it would be better to include such example data in a separate repository (e.g., with the example notebooks there instead of in docs/source/examples) or hosted on an s3 instance so they are downloaded (and maybe cached locally) once the users call the load_ methods of the dataset module.

openjournals/joss-reviews#1777
