### 2019-10-19 #1777 @martibosch
Hello @rmsare . Your TODO list seems a great response to my review. I will stay tuned.

Regarding the authorship, I did not intend to cast any doubt about the contributions made by second author, I just wanted to make sure that it is in line with the JOSS guidelines.
