### 2019-10-19 c2e57a3
Small formatting changes
