### 2019-10-20 49b340a
Remove data from package, download example data from S3 if needed
