+++ pymccrgb/datasets.py
@@ -2,10 +2,44 @@
 
 import os
 
+from urllib.request import urlretrieve
+
 from .ioutils import read_las
 
 path = os.path.dirname(__file__)
-DATA_DIRECTORY = os.path.join(path, "data")
+LOCAL_DATA_PATH = os.path.join(path, "data")
+REMOTE_DATA_URL = "https://pymccrgb-data.s3-us-west-2.amazonaws.com/"
+
+
+def load_dataset(filename, npoints=None):
+    """Loads sample dataset from an S3 bucket
+
+    Parameters
+    ----------
+        filename: str
+            Filename of remote/local file
+        npoints: int
+            Optional number of (random) points to load from file
+
+    Returns
+    -------
+        A data ndarray
+    """
+    if not filename.endswith((".las", ".laz")):
+        raise ValueError(
+            "Data file must be in LAS or LAZ format (.las, .laz)."
+            "You provided the file " + filename
+        )
+    if not os.path.exists(LOCAL_DATA_PATH):
+        os.mkdir(LOCAL_DATA_PATH)
+
+    local_filename = os.path.join(LOCAL_DATA_PATH, filename)
+    if not os.path.exists(local_filename):
+        url = os.path.join(REMOTE_DATA_URL, filename)
+        urlretrieve(url, local_filename)
+
+    data = read_las(local_filename, nrows=npoints)
+    return data
 
 
 def load_kaibab_lidar(npoints=None):
@@ -20,9 +54,9 @@ def load_kaibab_lidar(npoints=None):
     -------
         A data ndarray
     """
-    filename = os.path.join(DATA_DIRECTORY, "kaibab_lidar.laz")
-    data = read_las(filename, nrows=npoints)
-    return data
+    raise NotImplementedError("This sample dataset is not available yet!")
+    # data = load_dataset("kaibab_lidar.laz", npoints=npoints)
+    # return data
 
 
 def load_mammoth_lidar(npoints=None):
@@ -37,8 +71,7 @@ def load_mammoth_lidar(npoints=None):
     -------
         A data ndarray
     """
-    filename = os.path.join(DATA_DIRECTORY, "mammoth_lidar.laz")
-    data = read_las(filename, nrows=npoints)
+    data = load_dataset("mammoth_lidar.laz", npoints=npoints)
     return data
 
 
@@ -54,6 +87,6 @@ def load_mammoth_sfm(npoints=None):
     -------
         A data ndarray
     """
-    filename = os.path.join(DATA_DIRECTORY, "mammoth_sfm.laz")
-    data = read_las(filename, nrows=npoints)
-    return data
+    raise NotImplementedError("This sample dataset is not available yet!")
+    # data = load_dataset("mammoth_sfm.laz", npoints=npoints)
+    # return data
