### 2019-10-20 07e3fdd
Increase max-iter to ensure convergence
