### 2019-10-20 c427cff
Fix major linter errors and run Black
