+++ pymccrgb/__init__.py
@@ -1,3 +1,3 @@
 from . import core, datasets, features, ioutils, pointutils, plotting
 from .core import mcc, mcc_rgb
-from .ioutils import read_data 
+from .ioutils import read_data
