### 2019-10-20 c98bc71
Add black to CI checks
