### 2019-10-20 rmsare/pymccrgb#19 @rmsare
I've fixed the linter failures in the main codebase. I left tests and docs alone.

- Formatting uses black with a line length of 88 chars. I added this to CI.
- I chose to keep a couple of long lines where formatted strings are printed
- flake8 flags some of my convenience imports, like those in init.py and Axes3D. They are used, so  I didn't change these.
