### 2019-10-20 rmsare/pymccrgb#18 @rmsare
I decided to keep the docs as-is, but now download and cache example data from S3.

A separate docs/examples repo would be good if this grows into a larger project with more diverse examples, I agree. Usually I prefer to have the docs and example nbs included in the main repo - I feel this makes it easier to browse for examples and maintain the docs.
