### 2019-10-20 7852525
Try again with multiline script - use pipe
