### 2019-10-20 rmsare/pymccrgb#17 @rmsare
I am working on a [conda recipe](https://github.com/conda-forge/staged-recipes/pull/9923) for this package. When it's accepted, I will change the install instructions to something like:

*This package is developed for Linux/OS X and Python 3.6+. It depends on common Python packages like sklearn, numpy, the LibLAS C API, and MCC Python bindings.*

*You can install this package with conda:*

```
conda env create -n pymcc python=3.6
conda activate pymcc
conda install pymccrgb -y -c conda-forge
```
