### 2019-10-22 #1777 @kthyng
> As a reviewer, should I put a tick to the checklist anyway

@martibosch Yes please check it if all available doi's have been included.
