### 2019-10-22 d6964d3
Add license fileand bump verion to 0.1.5
