### 2019-10-23 #1777 @martibosch
Hello @rmsare @kthyng !

I believe @rmsare has addressed all my comments. I have checked all the boxes of the checklist. Everything is good from my side :tada:
