### 2019-10-23 #1777 @kthyng
Thanks @martibosch!

From what I can see, the reviewers comments have been addressed. It looks like @rmsare is working on getting the package into conda-forge, is that right?
