### 2019-10-25 933b1af
Update install instructions
