### 2019-10-26 98e3cb6
Remove unused crop function and fiona import
