### 2019-10-26 8434d94
Remove fiona dep
