+++ environment.yml
@@ -14,7 +14,6 @@ dependencies:
   - scikit-image
   - pdal
   - python-pdal
-  - fiona
   - pytest
   - pip:
     - git+https://github.com/stgl/pymcc
