### 2019-10-26 14efc9a
Install pymcc_lidar as conda package
