### 2019-10-26 999cd17
Remove unused deps and bump version
