### 2019-10-26 rmsare/pymccrgb#17 @rmsare
@martibosch The libLAS and wrapper conda recipes are ready to merge, but the main recipe is running into some fairly extensive dependency issues that will take time to resolve.

Is this a necessary step for your review, or could I work to resolve the conda packaging difficulties as the paper moves towards publication? Thanks.
