### 2019-10-28 rmsare/pymccrgb#17 @martibosch
Hello @rmsare! I understand that pushing recipes to conda-forge takes time and can easily get messy. I believe that all my other comments have been addressed and therefore, from my side, the paper can now move to publication (without having to wait for the conda recipe part).
Cheers!
