### 2019-10-28 rmsare/pymccrgb#17 @rmsare
Thanks! With a little debugging effort it will be up soon.
