### 2019-10-28 #1777 @rmsare
@kthyng The conda package is nearly there and I've made functional recipes to install the two major dependencies (libLAS and pymcc). However there are some conflicts that will take some time to fix.

Martí Bosch [agreed](https://github.com/rmsare/pymccrgb/issues/17#issuecomment-546835055) that we can proceed with the review without a finalized conda package. I am working to debug the recipe with conda-forge CI and hopefully this won't take more than a week or two to resolve.

I've archived the final version (0.1.6) at 10.5281/zenodo.3519582
