### 2019-10-29 97d8739
Update install instructions
