### 2019-10-29 #1777 @kthyng
@rmsare Ok makes sense. We can move forward then.

Please edit the metadata in your zenodo archive so that the title and author list match your paper exactly.

Also can I verify that v0.1.6 is the version that should be archived here?
