### 2019-10-29 #1777 @rmsare
Great, thank you! I updated the title and author list on Zenodo.

Yes, 0.1.6 is the version that should be archived with the paper.
