### 2019-10-29 #1777 @kthyng
@rmsare Can you fix the capitalization in your bibtex entries? To preserve capitalization from the .bib file to the resulting pdf, you can use {} around what you want to preserve.
