### 2019-10-30 764993d
Fix captialization in reference titles
