+++ paper/paper.bib
@@ -1,21 +1,21 @@
 @article{evans2007,
-  title     = "A Multiscale Curvature Algorithm for Classifying Discrete Return
-               {LiDAR} in Forested Environments",
-  author    = "Evans, J S and Hudak, A T",
-  journal   = "IEEE Trans. Geosci. Remote Sens.",
-  publisher = "ieeexplore.ieee.org",
+  title     = {A multiscale curvature algorithm for classifying discrete return
+               {LiDAR} in forested environments},
+  author    = {Evans, J S and Hudak, A T},
+  journal   = {IEEE Trans. Geosci. Remote Sens.},
+  publisher = {ieeexplore.ieee.org},
   volume    =  45,
   number    =  4,
-  pages     = "1029--1038",
+  pages     = {1029--1038},
   month     =  apr,
   year      =  2007,
   doi       =  {10.1109/TGRS.2006.890412}, 
 }
 
 @article{pedregosa2011,
-  title={Scikit-learn: Machine learning in Python},
+  title={Scikit-learn: Machine learning in {Python}},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
-  journal={Journal of machine learning research},
+  journal={Journal of Machine Learning Research},
   volume={12},
   number={Oct},
   pages={2825--2830},
@@ -26,7 +26,7 @@
 @article{vanderwalt2011, 
 author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}}, 
 journal={Computing in Science Engineering}, 
-title={The NumPy Array: A Structure for Efficient Numerical Computation}, 
+title={The {NumPy Array}: A structure for efficient numerical computation}, 
 year={2011}, 
 volume={13}, 
 number={2}, 
@@ -37,7 +37,7 @@ month={March},}
 
 @misc{pdal2018,
 author = {{PDAL Contributors}},
-title = {PDAL Point Data Abstraction Library},
+title = {{PDAL Point Data Abstraction Library}},
 year = 2018, 
 doi = {10.5281/zenodo.2556738}, 
 url = {https://doi.org/10.5281/zenodo.2556738}
