### 2019-10-30 #1777 @rmsare
@kthyng The title and journal capitalization should be fixed now
