### 2019-10-30 #1777 @kthyng
@rmsare Looks like it is borking on the scikit learn bib entry. You have an empty doi followed by a comma. The last line in a bib entry shouldn't have a comma, and perhaps you can leave out the doi line altogether since there isn't one? Then remove the comma from the line above it.

Also just know that I am ready to accept your submission once this goes through, so be sure it looks perfect to you.
