### 2019-10-30 #1777 @arfon
Yep, this PR should fix this: https://github.com/rmsare/pymccrgb/pull/21
