### 2019-10-31 ed07a64
Fix journal name with ampersand
