+++ paper/paper.bib
@@ -24,7 +24,7 @@
 
 @article{vanderwalt2011, 
 author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}}, 
-journal={Computing in Science Engineering}, 
+journal={Computing in Science and Engineering}, 
 title={The {NumPy Array}: A structure for efficient numerical computation}, 
 year={2011}, 
 volume={13}, 
