### 2019-10-31 af54888
Update with full scikit learn author list
