### 2019-10-31 #1777 @rmsare
Thank you for troubleshooting my Bibtex, and sorry for my delay - was travelling without web access.

I've accepted the empty doi PR and fixed two reference issues. The proof looks good to me.
