### 2019-10-31 #1777 @rmsare
Thank you for the editorial help and to @daniellivingston and @martibosch for their reviews.

I'll update/close the install issue once the conda package is accepted.
