### 2023-11-08 #1760 @sdruskat
> \<AT\>rmsare - it looks like the repo you entered doesn't exist - but it's just that the http:// was missing - I've added it now
> @arfon - does this need to be changed in an internal database as well?

Hi @arfon and @danielskatz, sorry to revive this after 4 years. I stumbled over the [live JSON representation](https://joss.theoj.org/papers/10.21105/joss.01777.json) of the metadata missing the protocol prefix for the `software_repository` field. I can catch this in my retrieval code, but should this be fixed in the DB?

![grafik](https://github.com/openjournals/joss-reviews/assets/3007126/93530a48-9dd3-497f-bb27-df54e5a8b3df)

Thanks!
